move user functionality to dedicated file
